+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](https://keepachangelog.com/), adapted for int
 
 ## [Unreleased]
 
+## [1229] - 2026-07-17
+
+- [TASK-808] Show Git commit stdout when stderr is empty
+
 ## [1228] - 2026-07-17
 
 - [TASK-806] Add a cost skill for task, objective, and project rollups